:truck: move theme to constants
